+++ src/controllers/ProductImageController.js
@@ -1,6 +1,7 @@
 const db = require('../config/db');
 
 exports.uploadImage = (req, res) => {
+  if (req.user.role_id !== 1) return res.status(403).json({ message: "Chỉ Admin mới có quyền xóa/thêm ảnh!" });
   const productId = req.params.productId;
 
   if (!req.file) return res.status(400).json({ message: 'Chưa chọn file ảnh' });
@@ -15,6 +16,7 @@ exports.uploadImage = (req, res) => {
   });
 };
 exports.deleteImage = (req, res) => {
+  if (req.user.role_id !== 1) return res.status(403).json({ message: "Chỉ Admin mới có quyền xóa/thêm ảnh!" });
   const imageId = req.params.imageId;
   
   const sql = "DELETE FROM product_images WHERE id = ?";
